Moved some more THS functionality around in the native library.
