+++ src/Native/LibTorchSharp/THSTensorFactories.cpp
@@ -439,3 +439,48 @@ Tensor THSTensor_kaiser_window(const int64_t len, bool periodic, double beta, co
 
     CATCH_TENSOR(torch::kaiser_window(len, periodic, beta, options));
 }
+
+double THSInit_calculate_gain(int64_t nonlinearity, double param)
+{
+    CATCH_RETURN(double, 0.0, torch::nn::init::calculate_gain(get_nl_type(nonlinearity), param))
+}
+
+torch::nn::init::FanModeType get_fan_mode(int64_t mode)
+{
+    return mode == 0 ? torch::nn::init::FanModeType(torch::kFanIn) : torch::nn::init::FanModeType(torch::kFanOut);
+}
+
+Tensor THSInit_constant_(Tensor tensor, Scalar value)
+{
+    CATCH_TENSOR(torch::nn::init::constant_(*tensor, *value))
+}
+
+Tensor THSInit_dirac_(Tensor tensor)
+{
+    CATCH_TENSOR(torch::nn::init::dirac_(*tensor))
+}
+
+Tensor THSInit_eye_(Tensor tensor)
+{
+    CATCH_TENSOR(torch::nn::init::eye_(*tensor))
+}
+
+Tensor THSInit_ones_(Tensor tensor)
+{
+    CATCH_TENSOR(torch::nn::init::ones_(*tensor))
+}
+
+Tensor THSInit_orthogonal_(Tensor tensor, double gain)
+{
+    CATCH_TENSOR(torch::nn::init::orthogonal_(*tensor, gain))
+}
+
+Tensor THSInit_sparse_(Tensor tensor, double sparsity, double std)
+{
+    CATCH_TENSOR(torch::nn::init::sparse_(*tensor, sparsity, std))
+}
+
+Tensor THSInit_zeros_(Tensor tensor)
+{
+    CATCH_TENSOR(torch::nn::init::zeros_(*tensor))
+}
